Add --existing-exposures argument to surveysim.
This argument allows the exposures file from the survey to be specified in addition to the tile file from the main survey.  When specified, the survey is simulated as if all existing exposures have already occurred.  i.e., the survey simulation is run starting at the current point in the survey.